Move test app under package 'com.adobe.edge.media.testapp'
