+++ code/testapp/src/main/java/com/adobe/edge/media/testapp/Configuration.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.mediaanalyticstestapp;
+package com.adobe.edge.media.testapp;
 
 public class Configuration {
