+++ code/testapp/src/main/java/com/adobe/edge/media/testapp/player/ObservableVideoView.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.mediaanalyticstestapp.player;
+package com.adobe.edge.media.testapp.player;
 
 import android.content.Context;
 import android.util.AttributeSet;
