+++ code/testapp/src/main/java/com/adobe/edge/media/testapp/player/PlayerEvent.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.mediaanalyticstestapp.player;
+package com.adobe.edge.media.testapp.player;
 
 public enum PlayerEvent {
     VIDEO_LOAD("video_load"),
