+++ code/testapp/src/main/java/com/adobe/edge/media/testapp/player/VideoPlayer.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.mediaanalyticstestapp.player;
+package com.adobe.edge.media.testapp.player;
 
 import android.app.Activity;
 import android.content.Context;
@@ -21,9 +21,9 @@ import android.os.HandlerThread;
 import android.os.Looper;
 import android.util.Log;
 import android.widget.MediaController;
+import com.adobe.edge.media.testapp.Configuration;
+import com.adobe.edge.media.testapp.R;
 import com.adobe.marketing.mobile.edge.media.MediaConstants;
-import com.adobe.mediaanalyticstestapp.Configuration;
-import com.adobe.mediaanalyticstestapp.R;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Observable;
